# openapi.yaml
openapi: 3.0.3
info:
  title: MyNotes API
  description: |
    REST API for the mynotes application. A `notes` resource demonstrates the
    full stack: OpenAPI → ogen server stubs → handler → service → repository →
    SQLite. Endpoints are served under the `/api/v1` base path (applied via the
    server URL and ogen's path prefix). Timestamps use RFC 3339 (UTC). Errors
    return `{"error": "message"}`.

    # Markdown dialect

    A note's `content` is stored and returned as verbatim Markdown. Consumers
    that render notes (the embedded web UI, the native mobile app, HTML export)
    should interpret it per the **Markdown dialect specification** in
    `markdown-spec.md`. In brief, it is [CommonMark](https://spec.commonmark.org/)
    plus a fixed subset of [GFM](https://github.github.com/gfm/) (tables,
    strikethrough, autolinks, task lists), a safe raw-HTML subset, and several
    application-specific render-time transforms — AsciiMath math (`$x^2$`), inline
    Lucide icons (`[!name]`), box / foldable blockquotes (`>-`, `>+`, `>*`) and
    the callouts built from them, internal wikilinks (`[[slug]]`), and artifact
    image references (`![alt](artifact:<sha256>)`). All
    transforms are stored verbatim; a consumer that passes `content` through
    unchanged receives the literal source. See `markdown-spec.md` for the
    full grammar, the callout alias/colour table, and the wikilink indexing rules.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
...
paths:
  /notes:
    get:
      operationId: listNotes
      summary: List notes
      description: >
        Returns a page of notes. The browse list (no `q`) is ordered by the
        `sort`/`order` parameters, defaulting to update time, newest first.
        When `q` is supplied, results are restricted to notes matching the
        full-text query and ordered by relevance (`sort`/`order` are ignored).
      parameters:
        - name: q
          in: query
          description: Optional full-text search query.
          required: false
          schema:
            type: string
            maxLength: 200
        - name: tag
          in: query
          description: >
            Optional tag slug filter. May be repeated; a note must carry ALL
            supplied tags to match (AND semantics). When combined with `q`, that
            filter also applies (AND). The single-value form (`tag=x`) is a
            one-element filter and behaves as before.
          required: false
          explode: true
          schema:
            type: array
            maxItems: 20
            items:
              type: string
              maxLength: 100
              pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        - name: titlePrefix
          in: query
          description: >
            When true and `q` is supplied, `q` is matched as a case-insensitive
            prefix of the note title (autocomplete style) instead of a full-text
            search; content is not matched. Ignored when `q` is empty.
          required: false
          schema:
            type: boolean
            default: false
        - name: sort
          in: query
          description: >
            Field the browse list is sorted by. `updated` (default) and
            `created` sort by the respective timestamp; `title` sorts
            alphabetically, case-insensitively. Ignored for full-text search
            (`q` without `titlePrefix`), which is ordered by relevance, and for
            `titlePrefix` matching, which is always ordered by title.
          required: false
          schema:
            type: string
            enum: [updated, created, title]
            default: updated
        - name: order
          in: query
          description: >
            Direction for `sort`. `desc` (default) is newest-first for the
            timestamp fields and Z→A for `title`; `asc` is the reverse.
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: A page of notes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteList"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createNote
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteRequest"
      responses:
        "201":
          description: The created note.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        default:
          $ref: "#/components/responses/Error"
...
components:
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Note:
      type: object
      required: [slug, title, content, created_at, updated_at, version, tags, incoming_links, outgoing_links]
      properties:
        slug:
          type: string
          maxLength: 100
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        title:
          type: string
          minLength: 1
          maxLength: 200
        content:
          type: string
          maxLength: 1000000
          description: >
            Verbatim Markdown content of the note. See the Markdown dialect
            specification (markdown-spec.md) for how consumers (web UI,
            native app) should interpret and render it.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        version:
          type: integer
          minimum: 1
          description: >
            Monotonically increasing edit counter. Starts at 1 on creation and
            increments by 1 on each write. No-op PATCHes do not increment it.
            Use with If-Match for optimistic locking.
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
          description: Tags attached to this note; empty array when none, never absent.
        incoming_links:
          type: array
          items:
            $ref: "#/components/schemas/NoteLink"
          description: >
            Backlinks: other notes whose content wikilinks to this note. Empty
            array when none, never absent. See NoteLink.
        outgoing_links:
          type: array
          items:
            $ref: "#/components/schemas/NoteLink"
          description: >
            Notes this note's content wikilinks to, resolved to existing notes
            only (dangling links to non-existent notes are omitted). Empty array
            when none, never absent. See NoteLink.
        published_at:
          type: string
          format: date-time
          description: >
            When the note was last published, if it is currently published;
            absent otherwise. The published page is a snapshot taken at this
            time, so a later `updated_at` means the public page is out of date.
    NoteSummary:
      type: object
      required: [slug, title, created_at, updated_at, excerpt, version, tags, incoming_links, outgoing_links]
      properties:
        slug:
          type: string
          maxLength: 100
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        title:
          type: string
          minLength: 1
          maxLength: 200
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        excerpt:
          type: string
          description: Short text excerpt; empty string when none, never absent.
        version:
          type: integer
          minimum: 1
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
          description: Tags attached to this note; empty array when none, never absent.
        incoming_links:
          type: array
          items:
            $ref: "#/components/schemas/NoteLink"
          description: >
            Backlinks: other notes whose content wikilinks to this note. Empty
            array when none, never absent. See NoteLink.
        outgoing_links:
          type: array
          items:
            $ref: "#/components/schemas/NoteLink"
          description: >
            Notes this note's content wikilinks to, resolved to existing notes
            only (dangling links to non-existent notes are omitted). Empty array
            when none, never absent. See NoteLink.
        published_at:
          type: string
          format: date-time
          description: >
            When the note was last published, if it is currently published;
            absent otherwise. See Note.published_at.
    NoteLink:
      type: object
      required: [slug, title]
      description: >
        A reference to another note in the wikilink graph: the target note's
        addressable slug and its current title. Only links whose target note
        exists are represented; tag links ([[#slug]]) are not part of this graph.
      properties:
        slug:
          type: string
          maxLength: 100
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        title:
          type: string
          minLength: 1
          maxLength: 200
    NoteList:
      type: object
      required: [total, notes]
      properties:
        total:
          type: integer
          description: Number of notes returned in this page.
        notes:
          type: array
          items:
            $ref: "#/components/schemas/NoteSummary"
    CreateNoteRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        content:
          type: string
          maxLength: 1000000
          default: ""
          description: >
            Markdown content; see the Markdown dialect specification
            (markdown-spec.md). HTML carrying disallowed constructs is
            rejected (400).
        slug:
          type: string
          maxLength: 100
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        tags:
          type: array
          items:
            type: string
            maxLength: 100
            pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          description: >
            Slugs of existing tags to attach. Tags must already exist (create
            via POST /tags first) — an unknown slug is a 400. Absent or
            omitted defaults to no tags.
...
    Tag:
      type: object
      required: [slug]
      properties:
        slug:
          type: string
          maxLength: 100
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string